Update modelsGen to translate an empty struct.
We've just added a new WebSocket message that's empty -- the specific
case in question is the MetricsRequest, which has no properties
(currently). This breaks the type translation, since no other types have
no properties.

This will translate an empty type to an "object", since the only
well-formed representation of an empty struct is an empty object, in our
case.

Signed-off-by: Paul Tagliamonte <paul@kittycad.io>